Move typecheck import logic to Root, use only absolute paths
